Add list check, update import/export, ignore own storage events
The list check also enables a more flexible URLs, as it adds
'/admindb' only if it is necessary. Also, your password is only
sent if we could verify that the URL is correct and points to a
valid Mailman installation.

Ignoring our own storage events helps eliminate strange behaviour
such as duplicate rendering or requests.